pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["jie <yanjya2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pandas = "^2.2.0"
google-generativeai = "^0.3.2"
langchain = "^0.1.5"
openai = "^1.11.1"
streamlit = "^1.31.0"
python-dotenv = "^1.0.1"
langchain-openai = "^0.0.5"
beautifulsoup4 = "^4.12.3"
chromadb = "^0.4.22"
fastapi = "^0.109.2"
python-telegram-bot = "^20.8"
uvicorn = "^0.27.0.post1"
requests = "^2.31.0"
pydantic = "^2.6.1"
pypdf = "^4.0.1"
faiss-cpu = "^1.7.4"
unstructured = "^0.12.4"
rank-bm25 = "^0.2.2"
line-bot-sdk = "^3.9.0"
flask = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
